Initial commit for server

+++ Cosmetics Dal/UsersDal.cs
@@ -2,30 +2,45 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Cosmetics_Dal
 {
     public class UsersDal : IUsersDal
     {
-        SHOPContext db;
+        private readonly SHOPContext db;
 
         public UsersDal(SHOPContext db)
         {
             this.db = db;
         }
+
         public void AddUsers(Users user)
         {
-            db.Users.Add(user);
-            db.SaveChanges();
+            var isExist = db.Users.Any(x => x.Username == user.Username);
+
+            if (!isExist)
+            {
+                db.Users.Add(user);
+                db.SaveChanges();
+            }
+            else
+            {
+                throw new Exception("Username already exists.");
+            }
         }
 
         public void DeleteUser(int userId)
         {
-            var user = db.Users.ToList().Find(x => x.UserId == userId);
-            db.Users.Remove(user);
-            db.SaveChanges();
+            var user = db.Users.Find(userId);
+            if (user != null)
+            {
+                db.Users.Remove(user);
+                db.SaveChanges();
+            }
+            else
+            {
+                throw new ArgumentException($"User with ID {userId} could not be found to delete.");
+            }
         }
 
         public List<Users> GetAllUsers()
@@ -35,14 +50,46 @@ namespace Cosmetics_Dal
 
         public Users GetUsersById(int id)
         {
-            return db.Users.ToList().Find(x => x.UserId == id);
+            return db.Users.Find(id);
+        }
+
+        // "Code" here refers to the user's identifying number (UserId)
+        public Users GetUserByCode(int userCode)
+        {
+            return db.Users.Find(userCode);
+        }
+
+        public Users GetUserByUsernameAndPassword(string username, string password)
+        {
+            return db.Users.FirstOrDefault(u => u.Username == username && u.Password == password);
         }
 
         public void UpdateUsers(Users user)
         {
-            var users = db.Users.ToList().Find(x => x.UserId == user.UserId);
-            users.Username = user.Username;
-            db.SaveChanges();
+            var existingUser = db.Users.Find(user.UserId);
+
+            if (existingUser != null)
+            {
+                // Update every editable field, not just Username
+                existingUser.Username = user.Username;
+                existingUser.FirstName = user.FirstName;
+                existingUser.LastName = user.LastName;
+                existingUser.Phone = user.Phone;
+                existingUser.Address = user.Address;
+                existingUser.Email = user.Email;
+
+                // Only overwrite the password if the caller actually supplied a new one
+                if (!string.IsNullOrEmpty(user.Password))
+                {
+                    existingUser.Password = user.Password;
+                }
+
+                db.SaveChanges();
+            }
+            else
+            {
+                throw new ArgumentException($"User with ID {user.UserId} could not be found to update.");
+            }
         }
     }
 }
